added label png conversion to matrix image

+++ convertPngToMatrix.py
@@ -5,7 +5,7 @@ import ntpath
 import os
 
 # make sure folder ends with a slash /
-folder = "/Users/lingevan/Workspace/SceneNet/29May/OutputSceneNet01good/"
+folder = "/Users/lingevan/Workspace/SceneNet/26May/OutputSceneNet01/"
 
 
 if not os.path.exists(folder+"depth/original"):
@@ -13,17 +13,18 @@ if not os.path.exists(folder+"depth/original"):
 if not os.path.exists(folder+"depth/proccessedVisualisation"):
     os.makedirs(folder+"depth/proccessedVisualisation")
 
+print 'processing depth'
 for image_path in glob.glob(folder+"depth/*.png"):
     imageName=ntpath.basename(image_path)
-    print 'processing depth ', imageName
-    plt.clf()
-    image = misc.imread(image_path)
-    
-    plt.imshow(image,interpolation='nearest')
-    plt.colorbar()    
-    plt.savefig(folder+'/depth/proccessedVisualisation/mat_'+imageName)
+    if not os.path.isfile(folder + '/proccessedVisualisation/mat_'+imageName):
+        plt.clf()
+        image = misc.imread(image_path)
+        
+        plt.imshow(image,interpolation='nearest')
+        plt.colorbar()    
+        plt.savefig(folder+'/depth/proccessedVisualisation/mat_'+imageName)
 
-    os.rename(image_path, folder+'/depth/original/'+imageName)
+        os.rename(image_path, folder+'/depth/original/'+imageName)
 
 
 if not os.path.exists(folder+"instance/original"):
@@ -31,17 +32,18 @@ if not os.path.exists(folder+"instance/original"):
 if not os.path.exists(folder+"instance/proccessedVisualisation"):
     os.makedirs(folder+"instance/proccessedVisualisation")
 
+print 'processing instance'
 for image_path in glob.glob(folder+"instance/*.png"):
     imageName=ntpath.basename(image_path)
-    print 'processing instance', imageName
-    plt.clf()
-    image = misc.imread(image_path)
-    
-    plt.imshow(image,interpolation='nearest')
-    plt.colorbar()    
-    plt.savefig(folder+'/instance/proccessedVisualisation/mat_'+imageName)
+    if not os.path.isfile(folder + '/proccessedVisualisation/mat_'+imageName):
+        plt.clf()
+        image = misc.imread(image_path)
+        
+        plt.imshow(image,interpolation='nearest')
+        plt.colorbar()    
+        plt.savefig(folder+'/instance/proccessedVisualisation/mat_'+imageName)
 
-    os.rename(image_path, folder+'/instance/original/'+imageName)
+        os.rename(image_path, folder+'/instance/original/'+imageName)
 
 
 if not os.path.exists(folder+"labels/original"):
@@ -49,16 +51,17 @@ if not os.path.exists(folder+"labels/original"):
 if not os.path.exists(folder+"labels/proccessedVisualisation"):
     os.makedirs(folder+"labels/proccessedVisualisation")
 
+print 'processing labels'
 for image_path in glob.glob(folder+"labels/*.png"):
     imageName=ntpath.basename(image_path)
-    print 'processing label', imageName
-    plt.clf()
-    image = misc.imread(image_path)
-    
-    plt.imshow(image,interpolation='nearest')
-    plt.colorbar()    
-    plt.savefig(folder+'/labels/proccessedVisualisation/mat_'+imageName)
+    if not os.path.isfile(folder + '/proccessedVisualisation/mat_'+imageName):
+        plt.clf()
+        image = misc.imread(image_path)
+        
+        plt.imshow(image,interpolation='nearest')
+        plt.colorbar()    
+        plt.savefig(folder+'/labels/proccessedVisualisation/mat_'+imageName)
 
-    os.rename(image_path, folder+'/labels/original/'+imageName)
+        os.rename(image_path, folder+'/labels/original/'+imageName)
